chore(deploy): updating scripts with useful utils
