+++ scripts/utils/index.js
@@ -1,7 +1,9 @@
 const { copyFolder } = require("./copyFolder");
 const { getInfo } = require("./info");
+const { stringifyInfoFile } = require("./modInfo");
 
 module.exports = {
 	copyFolder,
-	getInfo
+	getInfo,
+	stringifyInfoFile
 };
