+++ scripts/utils/modInfo.js
@@ -1,0 +1,14 @@
+/**
+ * Converts info object back to .info format
+ * @param {Record<string, string>}
+ * @returns {string}
+ */
+const stringifyInfoFile = info =>
+	Object.entries(info)
+		.filter(([, value]) => value !== undefined && value !== null)
+		.map(([key, value]) => `${key}=${value}`)
+		.join("\n");
+
+module.exports = {
+    stringifyInfoFile
+};
